server/sentiment/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
click==8.0.3
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
joblib==1.1.0
nltk==3.6.7
psycopg2==2.9.3
python-dotenv==0.19.2
pytz==2021.3
regex==2022.1.18
sqlparse==0.4.2
textblob==0.17.1
tqdm==4.62.3
